refactor(commands): reorganize command namespaces and update legacy imports
