+++ app/Console/Commands/legacy/LegacyAddTaxBudgetsCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\legacy;
 
 use App\Models\Legacy\LegacyBudgetPlan;
 use App\Models\TaxBudget;
