+++ app/Console/Commands/legacy/LegacyBudgetItemShift.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\legacy;
 
 use App\Models\Legacy\Booking;
 use App\Models\Legacy\LegacyBudgetItem;
