+++ app/Console/Commands/legacy/LegacyChangeBankAccountId.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\legacy;
 
 use App\Models\Legacy\BankAccount;
 use App\Models\Legacy\BankTransaction;
