+++ app/Console/Commands/legacy/LegacyMigrateFilesToStorage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\legacy;
 
 use App\Models\Legacy\ExpenseReceipt;
 use App\Models\Legacy\FileInfo;
