+++ app/Console/Commands/stufis/CheckEncryptedFieldsCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\stufis;
 
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Encryption\DecryptException;
