+++ app/Console/Commands/stufis/InvalidateBelegePdfCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\stufis;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
