+++ app/Console/Commands/stufis/StuFisHealth.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Console\Commands;
+namespace App\Console\Commands\stufis;
 
 use Illuminate\Console\Command;
